+++ azure-kusto-ingest/azure/kusto/ingest/_descriptors.py
