+++ azure-kusto-ingest/azure/kusto/ingest/_ingestion_properties.py
