+++ azure-kusto-ingest/azure/kusto/ingest/kusto_ingest_client.py
@@ -10,7 +10,7 @@ import tempfile
 from azure.storage.common import CloudStorageAccount
 
 from azure.kusto.data.request import KustoClient
-from .descriptors import BlobDescriptor, FileDescriptor
+from ._descriptors import BlobDescriptor, FileDescriptor
 from ._ingestion_blob_info import _IngestionBlobInfo
 from ._resource_manager import _ResourceManager
